+ test présence fichier avant suppression sinon erreur

+++ core/lib/class.plx.plugins.php
@@ -160,7 +160,8 @@ class plxPlugins {
 		elseif(isset($content['selection']) AND $content['selection']=='delete') {
 			foreach($content['chkAction'] as $idx => $plugName) {
 				if($this->deleteDir(realpath(PLX_PLUGINS.$plugName))) {
-					unlink(PLX_ROOT.PLX_CONFIG_PATH.'plugins/'.$plugName.'.xml');
+					if(is_file(PLX_ROOT.PLX_CONFIG_PATH.'plugins/'.$plugName.'.xml'))
+						unlink(PLX_ROOT.PLX_CONFIG_PATH.'plugins/'.$plugName.'.xml');
 					unset($this->aPlugins[$plugName]);
 				}
 			}
